version-a-durable-functions/requirements.txt
azure-functions
azure-functions-durable
requests